IMPORTANT COMMIT: Changed Coins system (see description)
Now coins aren't a class (it was pretty unuseful to make them a class, no sense), they're a struct in the scene and managed there too.

Also, implemented a clean up that deletes a PhysBody3D that you pass to that function.

Finally, added a way to delete a coin when picked. Now we can advance a lot :)

Co-Authored-By: jas3 <jaumeavinyo@users.noreply.github.com>